+++ includes/ach/class-woocommerce-ach-inovio-gateway.php
@@ -1,11 +1,11 @@
 <?php
 
 /**
- * It is use to create Ach payment method 
+ * It is use to create Ach payment method
  */
 include plugin_dir_path( __FILE__ ) . 'methods/class-ach-inovio-method.php';
 
-class Woocommerce_Ach_Inovio_Gateway extends Ach_Inovio_Method {
+class Woocommerce_Ach_Inovio_Rebill_Gateway extends Ach_Inovio_Method_Rebill {
 
     public function __construct() {
         parent::__construct();
@@ -13,4 +13,4 @@ class Woocommerce_Ach_Inovio_Gateway extends Ach_Inovio_Method {
 
 }
 
-new Woocommerce_Ach_Inovio_Gateway();
+new Woocommerce_Ach_Inovio_Rebill_Gateway();