+++ woocommerce-inovio-rebill-gateway.php
@@ -1,9 +1,9 @@
 <?php
 /**
- * Plugin Name: Zigu MX Payment Gateway
+ * Plugin Name: Zigu MX Payment Gateway Rebill
  * Description: Zigu MX payment gateway provide payment solutions.
  * Author: Zigu MX Payments
- * Version: 6.2.4
+ * Version: 6.2.5
  * Author URI: https://zigu.mx/
  * Plugin URI: https://www.zigu.mx
  * License: GPLv2
@@ -15,12 +15,12 @@ if ( !defined( 'ABSPATH' ) ) { // To check absolute path
     exit;
 }
 
-// Use to load Woocommerce_Inovio_init after plugin load
-add_action( 'plugins_loaded', 'Woocommerce_Inovio_init',11 );
+// Use to load Woocommerce_Inovio_Rebill_init after plugin load
+add_action( 'plugins_loaded', 'Woocommerce_Inovio_Rebill_init',11 );
 require plugin_dir_path( __FILE__ ) . 'includes/installer/inovio-plugin-database-table.php';
 // Create table for refund on activate inovio plugin
-register_activation_hook( __FILE__, 'create_inovio_plugin_database_table' );
-// register_activation_hook( __FILE__, 'create_ach_inovio_plugin_database_table' );
+register_activation_hook( __FILE__, 'create_inovio_rebill_plugin_database_table' );
+// register_activation_hook( __FILE__, 'create_ach_inovio_rebill_plugin_database_table' );
 require plugin_dir_path( __FILE__ ) . 'includes/common/class-common-inovio-payment.php';
 
 /**
@@ -28,22 +28,22 @@ require plugin_dir_path( __FILE__ ) . 'includes/common/class-common-inovio-payme
  *
  * @return null
  */
-function Woocommerce_Inovio_init() {
+function Woocommerce_Inovio_Rebill_init() {
 // Check class WP_Payment_Gateway exist or not
     if ( !class_exists( 'WC_Payment_Gateway' ) ) {
         return;
     }
 
 // Define Core Class Classes path
-    define( 'INOVIO_PLUGIN_CORE_CLASS', plugin_dir_path( __FILE__ ) . 'includes/common/inovio-core/' );
-    $class_files = scandir( INOVIO_PLUGIN_CORE_CLASS );
+    define( 'INOVIO_REBILL_PLUGIN_CORE_CLASS', plugin_dir_path( __FILE__ ) . 'includes/common/inovio-core/' );
+    $class_files = scandir( INOVIO_REBILL_PLUGIN_CORE_CLASS );
     foreach ( $class_files as $file ) {
         if ( preg_match( '/class-inovio?/', $file ) ) {
-            include_once INOVIO_PLUGIN_CORE_CLASS . $file;
+            include_once INOVIO_REBILL_PLUGIN_CORE_CLASS . $file;
         }
     }
 // hook to add link to inovio setting into checkout tab after activate inovio plugin
-    add_filter( 'plugin_action_links_' . plugin_basename( __FILE__ ), 'plugin_action_links' );
+    add_filter( 'plugin_action_links_' . plugin_basename( __FILE__ ), 'zigu_rebill_plugin_action_links' );
 
     /**
      * Add links to plugins page for checkout settings
@@ -51,7 +51,7 @@ function Woocommerce_Inovio_init() {
      * @param  array $links
      * @return array
      */
-    function plugin_action_links( $links = array() ) {
+    function zigu_rebill_plugin_action_links( $links = array() ) {
         $plugin_links = array (
             '<a href="' . admin_url( 'admin.php?page=wc-settings&tab=checkout' ) . '">' . __( 'Settings' ) . '</a>',
         );
@@ -65,20 +65,20 @@ function Woocommerce_Inovio_init() {
 }
 
 // Hook to add custom checkout field: woocommerce_review_order_before_submit
-    // add_action( 'woocommerce_review_order_before_submit', 'my_custom_checkout_field' );
+    // add_action( 'woocommerce_review_order_before_submit', 'zigu_rebill_custom_checkout_field' );
     /**
      * Add custom checkout field: woocommerce_review_order_before_submit
      *
      */
-    function my_custom_checkout_field() {
-        echo '<div id="my_custom_checkout_field">';
-            woocommerce_form_field( 'my_field_term', array (
+    function zigu_rebill_custom_checkout_field() {
+        echo '<div id="zigu_rebill_custom_checkout_field">';
+            woocommerce_form_field( 'zigu_rebill_field_term', array (
                 'type'      => 'checkbox',
                 'required'  => 'true',
                 'default' => 1,
                 'class'     => array ('input-checkbox'),
                 'label'     => __( 'I have read and agree to the <a href="'.plugin_dir_url( __FILE__ ) . 'assets/pdf/TOS_ENG.pdf'.'">Terms of Service</a> and <a href="'.plugin_dir_url( __FILE__ ) . 'assets/pdf/TOS_ENG.pdf'.'">Privacy Policy</a>'),
-            ),  WC()->checkout->get_value( 'my_field_term' ) );
+            ),  WC()->checkout->get_value( 'zigu_rebill_field_term' ) );
             echo '</div>';
         global $wpdb;
         $cart = WC()->cart->get_cart();
@@ -95,50 +95,49 @@ function Woocommerce_Inovio_init() {
                 $meta_interval = $meta['_subscription_period_interval'][0];
                 $regular_price = $meta['_subscription_price'][0];
                 $currency = get_woocommerce_currency_symbol();
-                echo '<div id="my_custom_checkout_field">';
-                    woocommerce_form_field( 'my_field_name', array (
+                echo '<div id="zigu_rebill_custom_checkout_field">';
+                    woocommerce_form_field( 'zigu_rebill_field_name', array (
                         'type'      => 'checkbox',
                         'required'  => 'true',
                         'default' => 1,
                         'class'     => array ('input-checkbox'),
                         'label'     => __('After '.$meta_interval.$meta_length.' , membership renews at '.$currency .$regular_price.' every '.$meta_interval.$meta_length.' untill cancelled'),
-                    ),  WC()->checkout->get_value( 'my_field_name' ) );
+                    ),  WC()->checkout->get_value( 'zigu_rebill_field_name' ) );
                     echo '</div>';
             }
         }
     }
 
 // Hook to save the custom checkout field in the order meta, when checkbox has been checked
-    add_action( 'woocommerce_checkout_update_order_meta', 'custom_checkout_field_update_order_meta', 10, 1 );
+    add_action( 'woocommerce_checkout_update_order_meta', 'zigu_rebill_checkout_field_update_order_meta', 10, 1 );
     /**
      * Save the custom checkout field in the order meta, when checkbox has been checked
      *
      * @param  int $order_id
      */
-    function custom_checkout_field_update_order_meta( $order_id ) {
-        if ( ! empty( sanitize_text_field( $_POST['my_field_name'] ) ) ) {
-            update_post_meta( $order_id, 'my_field_name', sanitize_text_field( $_POST['my_field_name'] ) );
+    function zigu_rebill_checkout_field_update_order_meta( $order_id ) {
+        if ( ! empty( sanitize_text_field( $_POST['zigu_rebill_field_name'] ) ) ) {
+            update_post_meta( $order_id, 'zigu_rebill_field_name', sanitize_text_field( $_POST['zigu_rebill_field_name'] ) );
         }
-        if ( ! empty( sanitize_text_field( $_POST['my_field_term'] ) ) ) {
-            update_post_meta( $order_id, 'my_field_term', sanitize_text_field( $_POST['my_field_term'] ) );
+        if ( ! empty( sanitize_text_field( $_POST['zigu_rebill_field_term'] ) ) ) {
+            update_post_meta( $order_id, 'zigu_rebill_field_term', sanitize_text_field( $_POST['zigu_rebill_field_term'] ) );
         }
     }
 
 // Hook to display the custom field result on the order edit page (backend) when checkbox has been checked
-    add_action( 'woocommerce_admin_order_data_after_billing_address', 'display_custom_field_on_order_edit_pages', 10, 1 );
+    add_action( 'woocommerce_admin_order_data_after_billing_address', 'zigu_rebill_display_custom_field_on_order_edit_pages', 10, 1 );
     /**
      * Display the custom field result on the order edit page (backend) when checkbox has been checked
      *
      * @param  array $order
      */
-    function display_custom_field_on_order_edit_pages( $order ) {
-        $my_field_name = get_post_meta( $order->get_id(), 'my_field_name', true );
-        $my_field_term = get_post_meta( $order->get_id(), 'my_field_term', true );
-        if( 1 == $my_field_name ) {
+    function zigu_rebill_display_custom_field_on_order_edit_pages( $order ) {
+        $zigu_rebill_field_name = get_post_meta( $order->get_id(), 'zigu_rebill_field_name', true );
+        $zigu_rebill_field_term = get_post_meta( $order->get_id(), 'zigu_rebill_field_term', true );
+        if( 1 == $zigu_rebill_field_name ) {
             echo '<p><strong>Product verification: </strong> <span style="color:red;">enabled</span></p>';
         }
-        if( 1 == $my_field_term ) {
+        if( 1 == $zigu_rebill_field_term ) {
             echo '<p><strong>Terms of Service: </strong> <span style="color:red;">enabled</span></p>';
         }
     }
-